Consolidate font face genenerators and other font related logic
